chore: rename /app/methods to /app/controllers
